Transfer to primary class files
