+++ include/kwpFrame.cpp
@@ -1,239 +1,133 @@
 #include "kwpFrame.h"
 
-/// @brief Parse essential metadata (target, sender, type) from a candidate CANFrame
-/// @param rxFrame Candidate CAN Frame that might be a KWP Frame
-void kwpFrame::parseMetaData(CanFrame* rxFrame)
+/// @brief Base constructor with direct assignment
+/// @param _target ID of the target ECU for this KWP Frame
+/// @param _sender ID of the sender ECU for this KWP Frame
+/// @param _SID SID of the KWP Frame
+/// @param _length Total payload length, including SID
+/// @param _bufferLength Total payload length, excluding SID
+/// @param _rxComplete Flag of completed reception (true by default)
+/// @param _multiFrame MultiFrame flag. False by default.
+kwpFrame::kwpFrame( byte _target, 
+                    byte _sender, 
+                    byte _SID, 
+                    uint16_t _length, 
+                    uint16_t _bufferLength, 
+                    bool _rxComplete, 
+                    bool _multiFrame) 
+                    :
+                    target(_target),
+                    sender(_sender),
+                    SID(_SID),
+                    length(_length),
+                    bufferLength(_bufferLength),
+                    rxComplete(_rxComplete),
+                    multiFrame(_multiFrame)
+{}
+
+/// @brief Base constructor with direct assignment
+/// @param _target ID of the target ECU for this KWP Frame
+/// @param _sender ID of the sender ECU for this KWP Frame
+/// @param _SID SID of the KWP Frame
+/// @param _dataBufLength Length of the base buffer to copy
+/// @param _dataBuf Base buffer to copy to the payload buffer, excluding SID
+/// @param _rxComplete Flag of completed reception (true by default)
+/// @param _multiFrame MultiFrame flag. False by default.
+kwpFrame::kwpFrame( byte _target, 
+                    byte _sender, 
+                    byte _SID, 
+                    uint16_t _dataBufLength, 
+                    const byte _dataBuf[], 
+                    bool _rxComplete,
+                    bool _multiFrame)
+                    :
+                    target(_target),
+                    sender(_sender),
+                    SID(_SID),
+                    length(_dataBufLength+1),
+                    bufferLength(_dataBufLength),
+                    rxComplete(_rxComplete),
+                    multiFrame(_multiFrame)
 {
-	sender = rxFrame->identifier & 0xFF;
-	target = rxFrame->data[0];
-	byte checkType = (rxFrame->data[1] & 0xF0);
-	if (checkType > 0x30 ) {
-		frameType = invalidFrameType;
-	}
-	else {
-		frameType = (KWP_FRAME_TYPE)checkType;
-	}
-	#ifdef KWP_FRAME_DEBUG
-		_debugSerial.printf("Ln%D\tFrame type parsed: %X\n",__LINE__,frameType);
-	#endif
+    for (int i = 0; i < _dataBufLength; i++)
+    {
+        buffer[i] = _dataBuf[i];
+    }
 }
 
-/// @brief 
-/// @param rxFrame 
-/// @return 
-KWP_FRAME_TYPE kwpFrame::processCanFrame(CanFrame *rxFrame)
+/// @brief Minimal constructor with target and sender ID
+/// @param _target ID of the target ECU for this KWP Frame
+/// @param _sender ID of the sender ECU for this KWP Frame
+kwpFrame::kwpFrame( byte _target, 
+                    byte _sender)
+                    :
+                    target(_target),
+                    sender(_sender)
 {
-	#ifdef KWP_FRAME_DEBUG
-		_debugSerial.printf("Ln%D\tProcessing CAN Frame to KWP Frame\n",__LINE__);
-	#endif
-	parseMetaData(rxFrame);
-	
-	switch(frameType) {
-		case singleFrame:
-			if(!RXComplete) {
-				#ifdef KWP_FRAME_DEBUG
-					_debugSerial.printf("Ln%D\tERROR: SingleFrame received while another frame is incomplete\n",__LINE__);
-				#endif  
-			}
-			else {
-				length          = rxFrame->data[1];
-				payloadLength   = length-1;
-				SID             = rxFrame->data[2];
-				cursor          = 0;
-
-				for(int i=3;i<3+payloadLength;i++) {
-					payload[cursor] = rxFrame->data[i];
-					cursor++;
-				} 
-			}
-			break;
-		case firstFrame:
-			if(!RXComplete){
-				#ifdef KWP_FRAME_DEBUG
-					_debugSerial.printf("Ln%D\tERROR: FirstFrame received while another frame is incomplete\n",__LINE__);
-				#endif  
-			} 
-			else {
-				length          = (swap_endian<uint16_t>(*((uint16_t*)&(rxFrame->data[1]))) & 0x0FFF);
-				payloadLength   = length -1;
-				SID             = rxFrame->data[3];
-				RXComplete      = false;
-				cursor          = 0;
-
-				for(int i = 4; i < 8;i++) {
-					payload[cursor] =   rxFrame->data[i];
-					cursor++;
-				}
-			}
-			break;
-		case continuationFrame:
-			if(!RXComplete) {
-				for(int i = 2; i<8; i++) {
-					if(cursor<payloadLength) {
-						payload[cursor] = rxFrame->data[i];
-						cursor++;
-					}
-					else {
-						RXComplete = true;
-						break;
-					}
-				}
-			}
-			else {
-				#ifdef KWP_FRAME_DEBUG
-					_debugSerial.printf("Ln%D\tERROR: ContinuationFrame without FirstFrame\n",__LINE__);
-				#endif
-			} 
-			break;
-		case flowControlFrame:
-			//Nothing to do, we are in the wrong scope for that
-			break;
-		default:
-			#ifdef KWP_FRAME_DEBUG
-				_debugSerial.printf("Ln%D\tERROR: Not a valid FrameType\n");
-			#endif
-			break;
-	}
-	return frameType;
 }
 
-/// @brief 
-/// @param singleShot When true, the frame broadcast will only be attempted once. False by default.
-/// @param loopBack When true, the CAN frames that are sent will be detected on the RX side. False by default.
-void kwpFrame::sendKwpFrame(bool singleShot ,bool loopBack)
+/// @brief Parses Metadata on first reception frame when in an RX position. Unused in TX
+/// @param canMetaFrame FirstFrame or SingleFrame (no Flow Control) to get the metadata from
+void kwpFrame::setMetadaData(twai_message_t *canMetaFrame)
 {
-	#ifdef KWP_FRAME_DEBUG
-		_debugSerial.printf("Ln%D\tSending KWP Frame ...\n",__LINE__);
-	#endif
-	
-	CanFrame txFrame;
-	txFrame.data_length_code = 8;
-	txFrame.identifier = (0x600 + sender);
-	txFrame.ss = (int)singleShot;
-	txFrame.self = (int)loopBack;
-	for (int i = 0; i < 8; i++)
-	{
-		txFrame.data[i] = 0xAA;
-	}
-	txFrame.data[0] = target;
-	#ifdef KWP_FRAME_DEBUG
-		_debugSerial.printf("Ln%D\t\tSTART : pendingFCFrame: %d RXComplete: %d TXComplete: %d Length: %d PayloadLength: %d Cursor: %d \n",__LINE__,pendingFCFrame,RXComplete,TXComplete,length,payloadLength,cursor);
-	#endif
-	if(!pendingFCFrame && TXComplete) { //Not pending a FCFrame and TX is complete
-		TXComplete = false;
-		if (length<=6) //SingleFrame is doable
-		{
-			#ifdef KWP_FRAME_DEBUG
-				_debugSerial.printf("Ln%D\t\tSingleFrame",__LINE__);
-			#endif
-			cursor = 0;
-			txFrame.data[1] = (byte)length;
-			txFrame.data[2] = SID;
-			for (int i = 0; i < payloadLength; i++)
-			{
-				txFrame.data[i+3] = payload[i];
-				cursor++;
-			}
-			ESP32Can.writeFrame(txFrame,TX_FRAME_TIMEOUT);
-			TXComplete = true;
-			#ifdef KWP_FRAME_DEBUG
-				_debugSerial.printf("\t SENT\n");
-			#endif
-		}
-		else //It is a first MF
-		{
-			#ifdef KWP_FRAME_DEBUG
-				_debugSerial.printf("Ln%D\t\tMultiFrame first part",__LINE__);
-			#endif
-			cursor = 0;
-			seqNumber = 1;
-			txFrame.data[1] = 0x10 + ((length & 0x0F00)>>8);
-			txFrame.data[2] = (length & 0xFF);
-			txFrame.data[3] = SID;
-			for (int i = 0; i < 4; i++)
-			{
-				txFrame.data[4+i]=payload[i];
-				cursor++;
-			}
-			ESP32Can.writeFrame(txFrame,TX_FRAME_TIMEOUT);
-			pendingFCFrame = true;
-			#ifdef KWP_FRAME_DEBUG
-				_debugSerial.printf("\t SENT\n");
-			#endif
-		}
-	}
-	else //Pending FC Frame... assumes this is in reaction to a FC frame and the transfer is incomplete
-	{
-		#ifdef KWP_FRAME_DEBUG
-			_debugSerial.printf("Ln%D\t\tContinuation frames\n",__LINE__);
-		#endif
-		while(!TXComplete) {
-			txFrame.data[1] = 0x20 + seqNumber;
-			for (int i = 0; i < 6; i++)
-			{
-				if(cursor<payloadLength) {
-					txFrame.data[2+i] = payload[cursor];
-					cursor++;
-				}
-				else
-				{
-					TXComplete = true;
-					txFrame.data[2+i] = 0xAA;
-				}
-			}
-			ESP32Can.writeFrame(txFrame,TX_FRAME_TIMEOUT);
-			#ifdef KWP_FRAME_DEBUG
-				_debugSerial.printf("Ln%D\t\t\tSENT with seq number 0x%02X\n",__LINE__,seqNumber);
-			#endif
-			seqNumber++;
-		}  
-		pendingFCFrame = false;
-	}
-	#ifdef KWP_FRAME_DEBUG
-		_debugSerial.printf("Ln%D\t\tFINAL : pendingFCFrame: %d RXComplete: %d TXComplete: %d Length: %d PayloadLength: %d Cursor: %d \n",__LINE__,pendingFCFrame,RXComplete,TXComplete,length,payloadLength,cursor);
-	#endif
+    target = canMetaFrame->data[0];
+    sender = canMetaFrame->identifier & 0xFF;
+    switch ((canMetaFrame->data[1]) & 0xF0)
+    {
+    case 0x00: //SingleFrame
+        multiFrame = false;
+        break;
+    case 0x10: //FirstFrame
+        multiFrame = true;
+        break;
+    default:
+        multiFrame = false;
+        break;
+    }
 }
 
-/// @brief 
-/// @param targetStream 
-void kwpFrame::printKwpFrame(Stream& targetStream)
+/// @brief After constructing a KWP Frame, calculates all metadata variants
+void kwpFrame::calculateMetaData()
 {
-	
-	if(RXComplete && (frameType != invalidFrameType && frameType != flowControlFrame)) {
-		targetStream.printf("Ln%D\t%02X | Sender: %02X Target: %02X SID: %02X PayloadLength: %3d |  ",__LINE__,frameType,sender,target,SID,payloadLength); 
-		for (int i = 0; i < payloadLength; i++)
-		{
-		targetStream.printf("%02X ",payload[i]);
-		}
-		targetStream.println();
-	}
-	else if (frameType == flowControlFrame)
-	{
-		targetStream.printf("FC TX: %02X RX: %02X\n",sender,target);
-	}
+    if(length>6) multiFrame = true;
+    else multiFrame = false;
 }
 
-/// @brief Used to reset Frames when in Timeout
-void kwpFrame::resetFrame()
+/// @brief Appends a segment of a CAN Frame buffer to an existing kwpFrame payload buffer
+/// @param canFrame Donor CAN Frame, referenced by pointer
+/// @param startPos Start position, by default 2
+/// @param endPos End position, by default 7
+/// @return True if the append is possible, false if sanity checks are failed
+bool kwpFrame::appendCanFrameBuffer(twai_message_t *canFrame, uint8_t startPos, uint8_t endPos)
 {
-	#ifdef KWP_FRAME_DEBUG
-		_debugSerial.printf("Ln%D\tResetting KWP Frame ...\n",__LINE__);
-	#endif
-	
-	target = 0x00;
-	sender = 0x00;
-	SID = 0x00;
-	length = 2;
-	payloadLength = 1;
-	cursor = 0;
-	RXComplete     =   true;
-	pendingFCFrame =   false;
-	TXComplete     =   true;
+    //Some sanity check
+    if( (startPos>endPos)   || 
+        (endPos>7)          || 
+        (canFrame==nullptr) || 
+        ((cursor + (endPos-startPos+1)) > bufferLength)) return false;
+    else
+    {
+        for (int i = startPos; i < endPos+1; i++)
+        {
+            buffer[cursor] = canFrame->data[i];
+            cursor++;
+        }
+        return true;
+    }
 }
 
-/// @brief 
-/// @param targetSerial 
-void kwpFrame::attachDebugSerial(Stream &targetSerial)
+/// @brief Resets frame to a fresh state
+void kwpFrame::reset(byte _target, byte _sender)
 {
-	_debugSerial = targetSerial;
+    target          =   _target;
+    sender          =   _sender;
+    SID             =   0x00;
+    length          =   2;
+    bufferLength   =   1;
+    cursor          =   0;
+    rxComplete      =   true;
+    for (int i = 0; i < 255; i++)
+    {
+        buffer[i] = 0x00;
+    }
 }
